add some benchmarks. change to uint8 from byte. use uint64 index

+++ sbloom.go
@@ -6,12 +6,12 @@ const (
 	idxMask  = 1<<elemSize - 1
 )
 
-func set(m []byte, n uint32) {
+func set(m []uint8, n uint64) {
 	idx, mask := n>>idxShift, n&idxMask
 	m[idx] |= 1 << mask
 }
 
-func get(m []byte, n uint32) bool {
+func get(m []uint8, n uint64) bool {
 	idx, mask := n>>idxShift, n&idxMask
 	return m[idx]|1<<mask != 0
 }
